框架：转换long主键 wojilu.apps

+++ wojilu.Apps/Blog/Domain/BlogApp.cs
@@ -17,7 +17,7 @@ namespace wojilu.Apps.Blog.Domain {
     [Serializable]
     public class BlogApp : ObjectBase<BlogApp>, IApp, IAccessStatus, IHits, ICommentApp {
 
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public String OwnerUrl { get; set; }
         public String OwnerType { get; set; }
