+++ wojilu.Apps/Content/Domain/ContentAttachment.cs
@@ -17,13 +17,13 @@ namespace wojilu.Apps.Content.Domain {
     [Serializable]
     public class ContentAttachment : ObjectBase<ContentAttachment>, ISort {
 
-        public int AppId { get; set; }
-        public int PostId { get; set; }
+        public long AppId { get; set; }
+        public long PostId { get; set; }
         
         [Column( Name="FileGuid", Length=40 )]
         public String Guid { get; set; }
 
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public String OwnerType { get; set; }
         public String OwnerUrl { get; set; }
