+++ wojilu.Apps/Content/Interface/IContentPostService.cs
@@ -14,7 +14,7 @@ namespace wojilu.Apps.Content.Interface {
 
     public interface IContentPostService {
 
-        ContentPost GetById( int postId, int ownerId );
+        ContentPost GetById(long postId, long ownerId);
 
         List<ContentPost> GetRelatedPosts( ContentPost post, int count );
         List<DataTagShip> GetRelatedDatas( ContentPost post, int count );
@@ -22,35 +22,35 @@ namespace wojilu.Apps.Content.Interface {
         ContentPost GetPrevPost( ContentPost post );
         ContentPost GetNextPost( ContentPost post );
 
-        List<ContentPost> GetRecentPost( int appId, int count, int typeId );
-        List<ContentPost> GetRankPost( int appId, int count, int typeId );
+        List<ContentPost> GetRecentPost(long appId, int count, int typeId);
+        List<ContentPost> GetRankPost(long appId, int count, int typeId);
 
-        int CountBySection( int sectionId );
-        List<ContentPost> GetAllBySection( int sectionId );
-        List<ContentPost> GetBySection( int sectionId );
-        List<ContentPost> GetBySection( int sectionId, int count );
-        List<ContentPost> GetTopBySectionAndCategory( int sectionId, int categoryId );
+        int CountBySection(long sectionId);
+        List<ContentPost> GetAllBySection(long sectionId);
+        List<ContentPost> GetBySection(long sectionId);
+        List<ContentPost> GetBySection(long sectionId, int count);
+        List<ContentPost> GetTopBySectionAndCategory(long sectionId, long categoryId);
 
-        DataPage<ContentPost> GetPageBySectionAndCategory( int sectionId, int categoryId );
-        DataPage<ContentPost> GetPageBySectionAndCategory( int sectionId, int categoryId, int pageSize );
+        DataPage<ContentPost> GetPageBySectionAndCategory(long sectionId, long categoryId);
+        DataPage<ContentPost> GetPageBySectionAndCategory(long sectionId, long categoryId, int pageSize);
 
-        DataPage<ContentPost> GetPageBySection( int sectionId );
-        DataPage<ContentPost> GetPageBySection( int sectionId, int pageSize );
+        DataPage<ContentPost> GetPageBySection(long sectionId);
+        DataPage<ContentPost> GetPageBySection(long sectionId, int pageSize);
 
         List<ContentPost> GetByIds( string ids );
-        List<ContentPost> GetByApp( int appId );
+        List<ContentPost> GetByApp(long appId);
 
-        DataPage<ContentPost> GetByApp( int appId, int pageSize );
+        DataPage<ContentPost> GetByApp(long appId, int pageSize);
 
-        DataPage<ContentPost> GetBySearch( int appId, string key, int pageSize );
+        DataPage<ContentPost> GetBySearch(long appId, string key, int pageSize);
         DataPage<ContentPost> GetTrashByApp( int appId, int pageSize );
 
-        ContentPost GetFirstPost( int appId, int sectionId );
+        ContentPost GetFirstPost(long appId, long sectionId);
 
-        DataPage<ContentPost> GetPageByCreator( int creatorId, IMember owner, int appId );
+        DataPage<ContentPost> GetPageByCreator(long creatorId, IMember owner, long appId);
 
-        int CountByCreator( int creatorId, IMember owner, int appId );
-        int CountByApp( int appId );
+        int CountByCreator(long creatorId, IMember owner, long appId);
+        int CountByApp(long appId);
 
 
         void AddHits( ContentPost post );
@@ -63,8 +63,8 @@ namespace wojilu.Apps.Content.Interface {
 
         void Delete( ContentPost post );
         void DeleteBatch( string ids );
-        void Restore( int id );
-        void DeleteSys( int postId );
+        void Restore(long id);
+        void DeleteSys(long postId);
         void DeleteTrueBatch( string ids );
 
         void UpdateAttachmentPermission( ContentPost post, int ischeck );
