+++ wojilu.Apps/Forum/Domain/ForumPost.cs
@@ -18,20 +18,20 @@ namespace wojilu.Apps.Forum.Domain {
     [Serializable]
     public class ForumPost : ObjectBase<ForumPost>, IPost, IAppData, IShareData, IHits, ILike {
 
-        public int AppId { get; set; }
-        public int ForumBoardId { get; set; }
+        public long AppId { get; set; }
+        public long ForumBoardId { get; set; }
 
-        public int TopicId { get; set; }
+        public long TopicId { get; set; }
 
         /// <summary>
         /// 只有主题本身对应的帖子，ParentId才是0
         /// </summary>
-        public int ParentId { get; set; }
+        public long ParentId { get; set; }
 
         public User Creator { get; set; }
         public String CreatorUrl { get; set; }
 
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public String OwnerType { get; set; }
         public String OwnerUrl { get; set; }
 
@@ -48,7 +48,7 @@ namespace wojilu.Apps.Forum.Domain {
 
         public int IsEdited { get; set; }
         public int EditCount { get; set; }
-        public int EditMemberId { get; set; }
+        public long EditMemberId { get; set; }
         public String EditReason { get; set; }
         public DateTime EditTime { get; set; }
